# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SynthChat Hermes Backend API
  version: 0.1.0
  description: |
    Frontend-facing contract for the complete Rust backend. Upstream Hermes
    HTTP, dashboard, WebSocket, CLI, filesystem and SQLite interfaces are
    compatibility references only and are not production runtime dependencies.
servers:
  - url: http://127.0.0.1:8642
    description: Local desktop backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/profiles/{profileId}/workspaces:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: List opaque workspace registrations
      responses:
        "200":
          description: Profile-scoped roots without local paths.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/SessionStorageUnavailable" }
    post:
      tags: [Workspaces]
      operationId: registerWorkspace
      summary: Register an existing absolute directory as a tool root
      description: The local path is accepted only in this write request, canonicalized by Rust and never returned by the API.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterWorkspace" }
      responses:
        "201":
          description: Workspace registered or idempotently replayed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/SessionStorageUnavailable" }
  /api/v1/profiles/{profileId}/workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/ProfileId"
      - $ref: "#/components/parameters/WorkspaceId"
    delete:
      tags: [Workspaces]
      operationId: deleteWorkspace
      summary: Delete an unreferenced workspace registration
      responses:
        "204": { description: Registration deleted; the directory and its contents are never modified. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/SessionStorageUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-desktop-session-token
  headers:
    ETag:
      description: Opaque resource revision, including quotes as required by HTTP.
      schema: { type: string }
  parameters:
    ProfileId: { name: profileId, in: path, required: true, schema: { $ref: "#/components/schemas/ProfileIdValue" } }
...
    WorkspaceId: { name: workspaceId, in: path, required: true, schema: { type: string, pattern: '^workspace_[A-Za-z0-9_]+$' } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for non-idempotent creates. Scoped by installation, method and canonical path; retained for at least 24 hours. Reuse with a different request fingerprint returns 409.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    BadRequest:
      description: Invalid request.
      x-error-codes: [validation_failed, invalid_if_match, invalid_cursor]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing or invalid desktop token.
      x-error-codes: [unauthorized]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      x-error-codes: [resource_not_found, profile_not_found, approval_not_found, clarification_not_found, hermes_state_not_found]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: State, revision, idempotency, replay-window or busy conflict.
      x-error-codes: [revision_conflict, idempotency_conflict, idempotency_resource_gone, mcp_server_exists, mcp_config_invalid, capacity_exceeded, unsafe_file_path, memory_storage_drift, profile_delete_conflict, session_busy, session_archived, event_history_expired, hermes_import_source_changed, hermes_import_conflict]
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    SessionStorageUnavailable:
      description: The local session database is unavailable or remained locked past its bounded busy timeout.
      x-error-codes: [session_storage_unavailable, session_storage_busy, hermes_state_unavailable]
      headers:
        Retry-After:
          description: Optional retry delay in seconds for a transient busy database.
          schema: { type: integer, minimum: 0 }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unprocessable:
      description: A required capability/provider is unavailable, or a semantically valid request violates a protected resource constraint.
      x-error-codes: [engine_capability_missing, memory_provider_unsupported, memory_content_blocked, memory_capacity_exceeded, session_search_unavailable, hermes_schema_unsupported, hermes_import_source_invalid, hermes_attachments_require_policy]
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    ProfileIdValue:
      type: string
      pattern: "^(default|[a-z0-9_][a-z0-9_-]{0,63})$"
...
    Workspace:
      type: object
      additionalProperties: false
      required: [id, profileId, displayName, available, createdAt, updatedAt]
      properties:
        id: { type: string, pattern: '^workspace_[A-Za-z0-9_]+$' }
        profileId: { $ref: "#/components/schemas/ProfileIdValue" }
        displayName: { type: string, minLength: 1, maxLength: 500 }
        available: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RegisterWorkspace:
      type: object
      additionalProperties: false
      required: [path]
      description: Desktop-local write-only path input. Responses and logs never echo this value.
      properties:
        path: { type: string, minLength: 1, maxLength: 16384 }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId, retryable]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"], format: uri-reference }
        code:
          type: string
          description: Stable machine-readable code. Each response component's x-error-codes lists known values without closing the set to future additions.
        requestId: { type: string }
        retryable: { type: boolean }